shim: Import from consolidated compat layer
