+++ lib/shim/cheerio.js
@@ -1,6 +1,6 @@
 /* global postman */
 
-import cheerio from '../cheerio.js'
+import { cheerio } from '../compat.js'
 
 const Extend = Symbol.for('extend')
