+++ lib/shim/crypto-js.js
@@ -1,7 +1,7 @@
 /* global postman */
 
-import cryptoJs from '../crypto-js.js'
+import compat from '../compat.js'
 
 const Extend = Symbol.for('extend')
 
-postman[Extend].module['crypto-js'] = cryptoJs
+postman[Extend].module['crypto-js'] = compat['crypto-js']
