+++ lib/shim/expect.js
@@ -1,6 +1,6 @@
 /* global postman */
 
-import chai from '../chai.js'
+import { chai } from '../compat.js'
 
 const Extend = Symbol.for('extend')
