+++ lib/shim/jsonSchema.js
@@ -1,6 +1,6 @@
 /* global postman */
 
-import Ajv from '../ajv.js'
+import { ajv as Ajv } from '../compat.js'
 
 const Extend = Symbol.for('extend')
