+++ lib/shim/lodash.js
@@ -1,6 +1,6 @@
 /* global postman */
 
-import lodash from '../lodash.js'
+import { lodash } from '../compat.js'
 
 const Extend = Symbol.for('extend')
