+++ lib/shim/urijs.js
@@ -1,6 +1,6 @@
 /* global postman */
 
-import URI from '../urijs.js'
+import { urijs as URI } from '../compat.js'
 
 const Extend = Symbol.for('extend')
