+++ lib/shim/xml2Json.js
@@ -1,4 +1,4 @@
-import xml2js from '../xml2js.js'
+import { xml2js } from '../compat.js'
 
 function xml2Json (xml) {
   let json
